refactor: split net standard PatchAsync implementation to another file
